tenstorrent: bh_arc: merge init.c and init_common.c
Merge init.c and init_common.c . Clean up unused header files.
