+++ lib/tenstorrent/bh_arc/eth.c
@@ -7,7 +7,7 @@
 #include "efuse.h"
 #include "eth.h"
 #include "harvesting.h"
-#include "init_common.h"
+#include "init.h"
 #include "noc.h"
 #include "noc2axi.h"
 #include "reg.h"
